md5: changed md5_update len param to uint32_t
This appears to be what the implementation is expecting. Also, the
original code used the unsigned int type.
This also fixes a vs8 warning message

+++ include/bmx/MD5.h
@@ -54,7 +54,7 @@ typedef struct
 
 
 void md5_init(MD5Context *ctx);
-void md5_update(MD5Context *ctx, const unsigned char *buf, size_t len);
+void md5_update(MD5Context *ctx, const unsigned char *buf, uint32_t len);
 void md5_final(unsigned char digest[16], MD5Context *ctx);
 
 std::string md5_digest_str(const unsigned char digest[16]);
